Add Config value and Refactor Harvest Manager to registry
Adds a config value which controls whether the player needs to use a tool to quick harvest. Also refactors the harvest manager to use a registry-style system, to protect the handler map, and open the system up for potential API. Also renames the Harvesters to Harvesters since it is a more sensible name.

Moves event handling to the HarvestManager, to allow for internal access to the classes stuff if that is required. Refactored the main mod class to get rid of redundant code.

Signed-off-by: Luke Gilfoyle <boo122333@gmail.com>