Adds better support for Berry Bushes
The problem: In vanilla berry bushes would drop their items on the ground, something which this mod avoids by placing the items directly in the player's inventory.

The solution: Extend the DefaultHarvester to accomadate plants that can be harvested before their max age, and add berry bushes to the harvesters using a DefaultHarvester.

Signed-off-by: Luke Gilfoyle (1700189) <l.gilfoyle@wlv.ac.uk>