Rewrite the mod
Adds new Config System, which streamlines the process of creating and adding a Config for a new Harvester.
Migrates the HarvestManager.java to a ForgeRegistry, as such we can now use the regsitry event and people who want to use the mod just need to listen to the RegistryEventHandler event for Harvesters.
Removes support for mushrooms, mushrooms are more like trees.
Adds support for berry bushes, like the berrybush branch.

Signed-off-by: Luke Gilfoyle (1700189) <l.gilfoyle@wlv.ac.uk>