Adds Harvester for WeepingVines and TwistingVines
Adds a harvester for weeping and twisting vines based on the kelp and tall plant harvesters.
Fixes a bug that existed silently for the kelp harvester and applied the logic to the tall plant harvester for continuity. The change was to use .equals when comparing the lowest point to the candidate position as to compare by contents rather than reference since it might be a different object.

Signed-off-by: Luke Gilfoyle <boo122333@gmail.com>